+++ tests/conftest.py
@@ -1,0 +1,3 @@
+import os
+
+os.environ.setdefault("LOKY_MAX_CPU_COUNT", "1")